refactor: folders-to-controllers-services-repos
