+++ src/auth/controllers/auth.controller.spec.ts
@@ -1,9 +1,9 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
+import { AuthService } from '../services/auth.service';
 import { validateOrReject } from 'class-validator';
-import { RegisterInput } from './dto/register.dto';
-import { LoginInput } from './dto/login.dto';
+import { RegisterInput } from '../dto/register.dto';
+import { LoginInput } from '../dto/login.dto';
 
 describe('AuthController', () => {
   let moduleRef: TestingModule;
