+++ src/auth/services/auth.service.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { JwtService } from '@nestjs/jwt';
 
 import { AuthService } from './auth.service';
-import { UserService } from '../user/user.service';
+import { UserService } from '../../user/services/user.service';
 
 describe('AuthService', () => {
   let service: AuthService;
