+++ src/auth/services/auth.service.ts
@@ -2,10 +2,10 @@ import { Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import { compare } from 'bcrypt';
 
-import { UserService } from '../user/user.service';
-import { User } from '../user/entities/user.entity';
-import { RegisterInput, RegisterOutput } from './dto/register.dto';
-import { LoginOutput } from './dto/login.dto';
+import { UserService } from '../../user/services/user.service';
+import { User } from '../../user/entities/user.entity';
+import { RegisterInput, RegisterOutput } from '../dto/register.dto';
+import { LoginOutput } from '../dto/login.dto';
 
 @Injectable()
 export class AuthService {
