+++ src/user/controllers/user.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { UserController } from './user.controller';
-import { UserService } from './user.service';
+import { UserService } from '../services/user.service';
 
 describe('UserController', () => {
   let controller: UserController;
