+++ src/user/user.module.ts
@@ -3,10 +3,10 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { SharedModule } from '../shared/shared.module';
 
-import { UserController } from './user.controller';
-import { UserService } from './user.service';
+import { UserController } from './controllers/user.controller';
+import { UserService } from './services/user.service';
 
-import { UserRepository } from './user.repository';
+import { UserRepository } from './repositories/user.repository';
 
 import { JwtStrategy } from '../auth/strategies/jwt.strategy';
